move all filter related config functions to filter config
