+++ src/deluge/model/mod_controllable/filters/filter_config.h
@@ -21,7 +21,35 @@
 #include "gui/l10n/strings.h"
 #include "util/misc.h"
 
-class Sound;
+// non namespaced
+enum class FilterMode {
+	TRANSISTOR_12DB,
+	TRANSISTOR_24DB,
+	TRANSISTOR_24DB_DRIVE, // filter logic relies on ladders being first and contiguous
+	SVF_BAND,              // first HPF mode
+	SVF_NOTCH,             // last LPF mode
+	HPLADDER,
+	OFF, // Keep last as a sentinel. Signifies that the filter is not on, used for filter reset logic
+};
+constexpr int32_t kNumFilterModes = util::to_underlying(FilterMode::OFF);
+constexpr FilterMode kLastLadder = FilterMode::TRANSISTOR_24DB_DRIVE;
+// Off is not an LPF mode but is used to reset filters
+constexpr int32_t kNumLPFModes = util::to_underlying(FilterMode::SVF_NOTCH) + 1;
+constexpr int32_t kFirstHPFMode = util::to_underlying(FilterMode::SVF_BAND);
+constexpr int32_t kNumHPFModes = util::to_underlying(FilterMode::OFF) - kFirstHPFMode;
+enum class FilterRoute {
+	HIGH_TO_LOW,
+	LOW_TO_HIGH,
+	PARALLEL,
+};
+
+constexpr int32_t kNumFilterRoutes = util::to_underlying(FilterRoute::PARALLEL) + 1;
+
+FilterRoute stringToFilterRoute(char const* string);
+char const* filterRouteToString(FilterRoute route);
+
+char const* lpfTypeToString(FilterMode lpfType);
+FilterMode stringToLPFType(char const* string);
 
 namespace deluge::dsp::filter {
